pyproject.toml
[project]
name = "project-recommendor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "invoke==2.2.0",
    "mongoengine==0.29.1",
    "pymongo==4.11.3",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "streamlit==1.44.0",
    "google-api-python-client==2.149.0",
    "google-auth==2.35.0",
]
